requirements/base.txt
Django==1.8.6
djangorestframework==3.5.3
django-cors-headers==1.3.1
